feat: bridge SQLCipher into Swift

+++ apple/IntegrationTests/KeychainProbe/Sources/main.swift
@@ -1,10 +1,12 @@
 import BondryApple
+import BondrySQLCipher
 import Darwin
 import Foundation
 import Security
 
 enum ProbeError: Error {
   case inconsistentKeys
+  case plaintextDatabase
   case cleanupFailed(OSStatus)
 }
 
@@ -31,6 +33,14 @@ do {
   let configuration = try KeychainDatabaseKeyConfiguration(service: service, account: account)
   let provider = KeychainDatabaseKeyProvider(configuration: configuration)
   let created = try provider.loadOrCreate()
+  let directory = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
+  try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
+  var needsDirectoryCleanup = true
+  defer {
+    if needsDirectoryCleanup {
+      try? FileManager.default.removeItem(at: directory)
+    }
+  }
   var needsCleanup = true
 
   defer {
@@ -42,6 +52,18 @@ do {
   guard try provider.load() == created, try provider.loadOrCreate() == created else {
     throw ProbeError.inconsistentKeys
   }
+
+  let databaseURL = directory.appendingPathComponent("bondry.db")
+  do {
+    let store = try BondryEncryptedStore.open(at: databaseURL, key: created)
+    try store.checkHealth()
+  }
+  let databaseBytes = try Data(contentsOf: databaseURL)
+  guard !databaseBytes.starts(with: Data("SQLite format 3\0".utf8)) else {
+    throw ProbeError.plaintextDatabase
+  }
+  try FileManager.default.removeItem(at: directory)
+  needsDirectoryCleanup = false
 
   let deleteStatus = deleteItem(service: service, account: account)
   needsCleanup = false
